requirements.txt
selenium
requests
beautifulsoup4
python-dotenv
gspread
google-auth
google-auth-oauthlib
